Fix linking inapps and subscriptions to userId

+++ java-tools/OsmAndServer/src/main/java/net/osmand/server/api/repo/DeviceSubscriptionsRepository.java
@@ -25,9 +25,9 @@ public interface DeviceSubscriptionsRepository extends JpaRepository<SupporterDe
 
 	List<SupporterDeviceSubscription> findFirst5BySkuOrderByStarttimeDesc(String sku);
 
-    List<SupporterDeviceSubscription> findAllByUserId(int userId);
+    List<SupporterDeviceSubscription> findAllByUserId(Integer userId);
 
-    List<SupporterDeviceSubscription> findAllBySupporterId(int supporterId);
+    List<SupporterDeviceSubscription> findAllBySupporterId(Long supporterId);
 
 	// PRIMARY KEY is (orderId + SKU) or (purchaseToken + SKU), orderId could be restored from purchaseToken and sku
 	@Entity
